Setup database instance and DI module & update kotlin to 2.1.0

+++ core/database/src/main/java/com/nrr/database/dao/TaskDao.kt
@@ -12,7 +12,7 @@ import kotlinx.coroutines.flow.Flow
 @Dao
 interface TaskDao {
     @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun insertTask(task: TaskEntity): Int
+    suspend fun insertTask(task: TaskEntity): Long
 
     @Transaction
     @Query("SELECT * FROM tasks")
